kotta/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "lxcor-kotta"
version = "0.1.0"
description = "Throttling and quota enforcement for Django REST Framework — per-endpoint rate limits, tier-based quotas, and usage tracking."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "lxcor", email = "i.alves.alex@gmail.com" }]
requires-python = ">=3.8"
keywords = ["django", "drf", "throttling", "rate-limiting", "quota"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "django>=3.2",
    "djangorestframework>=3.14",
    "lxcor-reggi>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/lxcor/api-suite/tree/main/kotta"
Repository = "https://github.com/lxcor/api-suite"

[tool.setuptools.packages.find]
where = ["."]
include = ["kotta*"]
